Fix BoardSizeText validation and progress bar updates
- Improved validation for BoardSizeText to handle empty/whitespace input.
- Fixed ProgressValueChanged event to ensure progress bar updates correctly.
- Adjusted throttling logic and event handling for progress updates.
- Resolved related test failures in validation and progress relay tests.

+++ NQueen.Domain/Interfaces/ISolverFrontEnd.cs
@@ -4,7 +4,7 @@ public interface ISolverFrontEnd
 {
     int DelayInMillisec { get; set; }
 
-    double ProgressValue { get; set; }
+    int ProgressValue { get; set; }
 
     event EventHandler<QueenPlacedEventArgs> QueenPlaced;
     event EventHandler<SolutionFoundEventArgs> SolutionFound;
